Move coachCard from classroom folder to coaches folder
